YG-160 🟢 게시글 수정 파일 분리
